requirements.txt
# Core Flask (unchanged)
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
gunicorn==23.0.0

# Database (unchanged)
alembic
pg8000
psycopg2-binary
SQLAlchemy

# DeepFace stack (CPU-ONLY VERSIONS - MUCH SMALLER)
tensorflow-cpu==2.13.0
deepface==0.0.79
opencv-python-headless==4.8.0.74
numpy==1.24.3

# Utilities (unchanged)
bcrypt
pillow
greenlet

# PyTorch CPU-only (if you really need it)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.0+cpu
torchvision==0.15.0+cpu
psutil==5.8.0  # For system resource monitoring